checking the search term looks like an accession before checking the utils/accession
