+++ package.json
@@ -16,7 +16,7 @@
     "test": "npm-run-all test:flow test:lint test:unit",
     "test:flow": "flow focus-check --merge-timeout=1000 .",
     "test:lint": "eslint --color --ignore-path .gitignore --ignore-pattern \"flow-typed\" --ignore-pattern \"snippets\" \"./**/*.js\"",
-    "test:unit": "jest --notify --maxWorkers=2 src",
+    "test:unit": "jest --maxWorkers=2 src",
     "test:functional": "[[ -n $IS_EBI_CLUSTER ]] && echo 'skipping functional tests on EBI cluster' || echo 'NOTICE: This option requires Ipro7 to be released in dist/' && jest --detectOpenHandles --config jest.config.js --notify tests",
     "test:functional:dev": "[[ -n $IS_EBI_CLUSTER ]] && echo 'skipping functional tests on EBI cluster' || echo 'NOTICE: Running tests against dev server' && jest --detectOpenHandles --config jest.config.js --testURL=https://wwwdev.ebi.ac.uk/interpro --notify tests",
     "test:functional:beta": "[[ -n $IS_EBI_CLUSTER ]] && echo 'skipping functional tests on EBI cluster' || echo 'NOTICE: Running tests against beta server' && jest --detectOpenHandles --config jest.config.js --testURL=https://www.ebi.ac.uk/interpro/beta --notify tests",
